Make the asyncToSync context refcounted
If the worker thread takes too long and the cond wait times out (e.g. if
the network interface is down) then the main thread will clean up the
context. The result is that the worker thread will eventually time out
from CURL and then try to do a cond_broadcast on something that no
longer exists, resulting in the hang.