Fix-Cmake: installation of entities public headers
(Hereinbelow comment associated with tests on RHEL8.3 OS)

The installation of libredfish did not copy two headers required/included by
redfish.h, namely entities/chassis.h and entities/resource.h.
As a result, it was impossible to compile C code with a dependency on
libredfish.

The proposed fix is quick and simple, not to say quick and dirty.
A list of public headers to install was created specifically for the
entities subfolder (named REDFISH_HDR_PUBLIC_ENTITIES).
For consistency, the initial list of public header, namely REDFISH_HDR_PUBLIC,
was renamed REDFISH_HDR_PUBLIC_RED (since the names of all the header belonging
to it are prefixed by "red").