Fix-Cmake: missing dependency at link edition for redfishcli
(Hereinbelow comment associated with tests on RHEL8.3)

On RHEL8, the readline-devel RPM is not available (even when EPEL8 is taken
into account).
As a result, it is compulsory to compile and install readline from source.

At compile time (i.e while running make after the Makefile was generated
through the action of cmake), it appears that some symbols used by
libreadline are not defined:

/usr/local/lib/libreadline.so: undefined reference to 'tgetstr'
/usr/local/lib/libreadline.so: undefined reference to 'tputs'
/usr/local/lib/libreadline.so: undefined reference to 'BC'
/usr/local/lib/libreadline.so: undefined reference to 'tgetent'
/usr/local/lib/libreadline.so: undefined reference to 'tgetflag'
/usr/local/lib/libreadline.so: undefined reference to 'tgoto'
/usr/local/lib/libreadline.so: undefined reference to 'UP'
/usr/local/lib/libreadline.so: undefined reference to 'tgetnum'
/usr/local/lib/libreadline.so: undefined reference to 'PC'

Those symbols are defined by libtinfo, which can be installed from the official
RHEL8 repositories.
As a result, to be able to build redfishcli, which depends on libreadline, it
seems that it is required to also link on libtinfo.

To do so, FindReadline.cmake was modified so as to look for the path to
libtinfo, and return it into ${Tinfo_LIBRARY}.
CMakeLists.txt was then modified to link to ${Tinfo_LIBRARY} when building
redfishcli.