Fix out-of-bounds string manipulation causing segfault.
If registerForEvents() is called with "libredfish" for the postbackUri
parameter, the function accepts this string and passes a pointer to
'postbackUri+11' to getDestinationAddress. In this case, the pointer actually
points past the end of the string's null byte.
On CHERI architectures, such as ARM Morello, pointer bounds are enforced in
hardware and attempting to dereference the pointer passed to
getDestinationAddress() causes a segfault.

Valid values for postbackUri should include a colon after "libredfish",
checking for this as part of the strncmp call rejects the invalid string
"libredfish" and this also means that getDestinationAddress() is not passed
an invalid pointer. This prevents a segfault on CHERI and prevents undefined
behavior on other architectures.

Signed-off-by: Michael Cobb <michael.cobb@iceotope.com>